backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
Pillow==11.0.0
requests==2.32.3
python-dotenv==1.0.1
google-genai==0.7.0
opencv-python-headless==4.10.0.84
numpy>=1.26.0
scipy>=1.11.0
scikit-image==0.22.0
python-multipart
google-cloud-aiplatform>=1.66.0


# Optional for even better results
 albumentations==1.3.1
